Update/versions (#10)
* [upd] replace 'scan' with 'scout cves'
* [upd] versions updated
* [workaround] ubuntu ssl workaround
* [workaround] disable .net globalization (not supported on 23.04)

Updated to the latest (available/supported) versions of:

ubuntu (23.04) [base layer updated]
golang 1.20.4
go-junit-report 2.0.0
go-swagger 0.30.4
maven 3.9.2
helm 3.12.0
nvm 0.39.3
git 2.39.2
docker (cli) 20.10.24
node 16.20.0

The Azure DevOps Agent is unable to start on ubuntu 23.04 due to some issue with the ICU and Globalization support not working with .net core 3.1. This may be addressed in the 3.x Agent software which uses .net 6.0, but in the meantime the solution is to disable .net core Globalization.